refactor(claude): extract question-answer input builders to a shared util
Move the AskUserQuestion / request_user_input answer-to-input builders (and the
question tool-name predicates) out of the SDK permission handler into a standalone
util. No behavior change — the SDK handler imports them — so the PTY permission
bridge can reuse them without pulling in the SDK handler's dependencies.
